+++ tests/testthat/test_pipelines.R
@@ -1,29 +1,6 @@
 task = tsk("rats")$filter(sample(300, 50L))
 task_regr = tgen("friedman1")$generate(20L)
 
-test_that("crankcompositor", {
-  pipe = mlr3pipelines::ppl("crankcompositor", learner = lrn("surv.kaplan"))
-  expect_class(pipe, "Graph")
-  pipe = mlr3pipelines::ppl("crankcompositor", learner = lrn("surv.kaplan"), graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  pipe$train(task)
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-  expect_true("crank" %in% p$predict_types)
-})
-
-test_that("distrcompositor", {
-  pipe = mlr3pipelines::ppl("distrcompositor", learner = lrn("surv.rpart"))
-  expect_class(pipe, "Graph")
-
-  pipe = mlr3pipelines::ppl("distrcompositor", learner = lrn("surv.rpart"), graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-
-  p = pipe$train(task)$predict(task)
-  expect_prediction_surv(p)
-  expect_true("distr" %in% p$predict_types)
-})
-
 test_that("survaverager", {
   pipe = mlr3pipelines::ppl("survaverager", learners = list(lrn("surv.kaplan"),
     lrn("surv.kaplan", id = "k2")))
@@ -45,67 +22,6 @@ test_that("survbagging", {
   pipe$train(task)
   p = pipe$predict(task)
   expect_prediction_surv(p)
-})
-
-test_that("resample survtoregr", {
-  pipe = mlr3pipelines::ppl("survtoregr", method = 1, distrcompose = FALSE, graph_learner = TRUE)
-  rr = resample(task, pipe, rsmp("cv", folds = 2L))
-  expect_numeric(rr$aggregate())
-})
-
-test_that("survtoregr 1", {
-  pipe = mlr3pipelines::ppl("survtoregr", method = 1, distrcompose = FALSE)
-  expect_class(pipe, "Graph")
-  pipe = mlr3pipelines::ppl("survtoregr", method = 1, distrcompose = FALSE, graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  pipe$train(task)
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-
-  pipe = mlr3pipelines::ppl("survtoregr", method = 1, distrcompose = TRUE, graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  pipe$train(task)
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-  expect_true("distr" %in% p$predict_types)
-})
-
-test_that("survtoregr 2", {
-  pipe = mlr3pipelines::ppl("survtoregr", method = 2)
-  expect_class(pipe, "Graph")
-  pipe = mlr3pipelines::ppl("survtoregr", method = 2, graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  pipe$train(task)
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-  expect_true("distr" %in% p$predict_types)
-
-  pipe = mlr3pipelines::ppl("survtoregr", method = 2, regr_se_learner = lrn("regr.featureless"),
-    graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  pipe$train(task)
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-  expect_true("distr" %in% p$predict_types)
-})
-
-test_that("survtoregr 3", {
-  pipe = mlr3pipelines::ppl("survtoregr", method = 3, distrcompose = FALSE)
-  expect_class(pipe, "Graph")
-  pipe = mlr3pipelines::ppl("survtoregr", method = 3, distrcompose = FALSE,
-                            graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  suppressWarnings(pipe$train(task)) # suppress loglik warning
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-
-  pipe = mlr3pipelines::ppl("survtoregr", method = 3, distrcompose = TRUE,
-                            graph_learner = TRUE)
-  expect_class(pipe, "GraphLearner")
-  suppressWarnings(pipe$train(task)) # suppress loglik warning
-  p = pipe$predict(task)
-  expect_prediction_surv(p)
-  expect_true("distr" %in% p$predict_types)
 })
 
 skip_if_not_installed("mlr3learners")